Separate config part of config-sc
